Added FieldInfo, FieldPosition and RobotPosition classes.
FieldInfo class will contain information about field elements and
starting positions, etc. FieldPosition represents the position of a
point on the field. RobotPosition represents a position and a direction
of a robot on the field. Coordinates are x and y with +X closer to the
judges than -X and +Y closer to the blue alliance station than -Y.
Directions are in degrees with 0.0 degrees facing the judges.